4.21

+++ part4/bloglist/controllers/blogs.js
@@ -7,21 +7,12 @@ blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({}).populate('user')
   res.json(blogs.map(blog => blog.toJSON()))
 })
-
-const getTokenFrom = req => {
-  const authorization = req.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
   
-blogsRouter.post('/', async (req, res, next) => {
+blogsRouter.post('/', async (req, res) => {
   const body = req.body
 
-  const token = getTokenFrom(req)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!req.token || !decodedToken.id) {
     return res.status(401).json({ error: 'token missing or invalid' })
   }
   const user = await User.findById(decodedToken.id)
@@ -65,6 +56,19 @@ blogsRouter.put('/:id', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res, next) => {
+  const body = req.body
+
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+  const userId = decodedToken.id
+
+  const blog = await Blog.findById(req.params.id)
+  if (blog.user.toString() !== userId.toString()) {
+    return res.status(401).json({ error: 'permission denied '})
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
